Fix Qt6CmeraWidget
- Add hidden files to .gitignore
- Fix bug that default resolution is (-1 x -1)
- Fix Qt6CameraWidget to remove redundant resolutions. Previously it will give multiple menuitems on UI with the same resolutions because internally the available camera formats include several with same resolutions but different pixel formats. This PR removes those redundant ones by simply preseving only the first formats with the same resolutions when initializing Qt6CameraWidget.
- Remove unused var Qt6CameraWidget::m_resolution.

Note: the resulting resolutions provided to user may mix different pixel formats. When changing the camera resolution with different pixel formats on the program, it may corrupt the camera video feed. This is not a big concern because it so far only occurs on not-used resolutions like 720 x 576, and the corrupt video feed can be fixed by changing to a new resolution.