Update EggFetcher_Core.c
Extended the delays in the fly to gate function to combat lag in pokemn SV introduced post the new updates. Also made it so when the program starts the program first automatically flies to the gate inorder to combat any positioning error that may have been caused due to stick drift. Also makes it simpler to launch the program as now the user just needs to be in the crater as opposed to earlier where they had to fly to it and stand on roughly the exact same spot.