Update EggHatcher_Core.c
running it for 3 consective days straight showed that after a while the game starts to lag while closing the boxes when getting the eggs, intefering with closing it properly and subsequently causing the loop to not function properly resulting in unhatched eggs. A simple wait function before closing the boxes seems to fix it for the most part.