Added error checking to filter menu
Added error checking to filter menu so it can leave properly, was previously crashing because it was stuck when leaving Pokemon summary was slow. Still error checking to do on the 4 positions the cursor can be on when filtering, some options become greyed out and will not accept an A button press.